requirements.txt
dnspython==2.2.1
numpy==1.23.5
pandas==1.5.2
pyarrow==10.0.1
pymongo==4.3.3
pymongoarrow==0.6.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
